ACE driver rebuilt with 'Arm GNU Toolchain 12.2.Rel1'
this is the same version of the toolchain as we're using to build the bB
custom code. there is either a bug in the older toolchain 'GNU Arm
Embedded Toolchain 10.3-2021.10' (unlikely) or there's issue caused with
how the two toolchains are interacting

note that we have to build the custom code with the newer toolchain
because we need access to the -Oz optimisation flag. That doesn't exist
in the the older version of the GCC

The version number of the ACE driver has not been changed because the
underlying code has not altered. The MD5 sum as reported by the -check
option will of course be different